fastapi-social-sentiment/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
transformers==4.40.1
python-dotenv==1.0.1
requests==2.31.0
elasticsearch==8.8.1